knowledgebase_backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==24.1.1
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
# uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
